+++ pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>dev.sefiraat</groupId>
-    <artifactId>BotanyStory</artifactId>
+    <artifactId>Cultivation</artifactId>
     <version>MODIFIED</version>
 
     <build>
@@ -19,19 +19,19 @@
                     <relocations>
                         <relocation>
                             <pattern>org.bstats</pattern>
-                            <shadedPattern>dev.sefiraat.botanystory.bstats</shadedPattern>
+                            <shadedPattern>dev.sefiraat.cultivation.bstats</shadedPattern>
                         </relocation>
                         <relocation>
                             <pattern>io.github.bakedlibs.dough</pattern>
-                            <shadedPattern>dev.sefiraat.botanystory.dough</shadedPattern>
+                            <shadedPattern>dev.sefiraat.cultivation.dough</shadedPattern>
                         </relocation>
                         <relocation>
                             <pattern>io.papermc.lib</pattern>
-                            <shadedPattern>dev.sefiraat.botanystory.paperlib</shadedPattern>
+                            <shadedPattern>dev.sefiraat.cultivation.paperlib</shadedPattern>
                         </relocation>
                         <relocation>
                             <pattern>io.github.sefiraat.sefilib</pattern>
-                            <shadedPattern>io.github.sefiraat.botanystory.sefilib</shadedPattern>
+                            <shadedPattern>io.github.sefiraat.cultivation.sefilib</shadedPattern>
                         </relocation>
                     </relocations>
                     <filters>
@@ -81,7 +81,7 @@
         <maven.compiler.source>16</maven.compiler.source>
         <maven.compiler.target>16</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <sonar.projectKey>BotanyStory</sonar.projectKey>
+        <sonar.projectKey>Cultivation</sonar.projectKey>
         <sonar.organization>sefiraat</sonar.organization>
         <sonar.host.url>https://sonarcloud.io</sonar.host.url>
