--- src/main/java/dev/sefiraat/botanystory/api/interfaces/CustomPlacementBlock.java
@@ -1,11 +1,0 @@
-package dev.sefiraat.botanystory.api.interfaces;
-
-import org.bukkit.event.block.BlockPlaceEvent;
-
-import javax.annotation.Nonnull;
-
-public interface CustomPlacementBlock {
-
-    void whenPlaced(@Nonnull BlockPlaceEvent event);
-
-}
