--- src/main/java/dev/sefiraat/botanystory/api/utils/LevelType.java
@@ -1,16 +1,0 @@
-package dev.sefiraat.botanystory.api.utils;
-
-public enum LevelType {
-    /**
-     * Plants
-     */
-    HORTICULTURALIST,
-    /**
-     * Bushes
-     */
-    TOPIARIST,
-    /**
-     * Trees
-     */
-    ORCHARDIST
-}
