--- src/main/java/dev/sefiraat/botanystory/managers/SupportedPluginManager.java
@@ -1,25 +1,0 @@
-package dev.sefiraat.botanystory.managers;
-
-import com.google.common.base.Preconditions;
-
-/**
- * This class is used to check, register and manage/provide supported plugins
- */
-public class SupportedPluginManager {
-
-    private static SupportedPluginManager instance;
-
-    public SupportedPluginManager() {
-        Preconditions.checkArgument(instance == null, "Cannot create a new instance of the SupportedPluginManager");
-        instance = this;
-        setup();
-    }
-
-    private void setup() {
-
-    }
-
-    public static SupportedPluginManager getInstance() {
-        return instance;
-    }
-}
