+++ src/main/java/dev/sefiraat/cultivation/api/interfaces/CustomPlacementBlock.java
@@ -1,0 +1,11 @@
+package dev.sefiraat.cultivation.api.interfaces;
+
+import org.bukkit.event.block.BlockPlaceEvent;
+
+import javax.annotation.Nonnull;
+
+public interface CustomPlacementBlock {
+
+    void whenPlaced(@Nonnull BlockPlaceEvent event);
+
+}
