+++ src/main/java/dev/sefiraat/cultivation/api/utils/LevelType.java
@@ -1,0 +1,16 @@
+package dev.sefiraat.cultivation.api.utils;
+
+public enum LevelType {
+    /**
+     * Plants
+     */
+    HORTICULTURALIST,
+    /**
+     * Bushes
+     */
+    TOPIARIST,
+    /**
+     * Trees
+     */
+    ORCHARDIST
+}
