+++ src/main/java/dev/sefiraat/cultivation/implementation/utils/Keys.java
@@ -1,0 +1,24 @@
+package dev.sefiraat.cultivation.implementation.utils;
+
+import dev.sefiraat.cultivation.Cultivation;
+import org.bukkit.NamespacedKey;
+
+import javax.annotation.Nonnull;
+
+public final class Keys {
+
+    private Keys() {
+        throw new IllegalStateException("Utility class");
+    }
+
+    // Region Slimefun BlockStorage Keys
+    public static final String SEED_GROWTH_STAGE = "growth-stage";
+    public static final String BLOCK_OWNER = "plant-owner";
+    public static final NamespacedKey COOLDOWN = newKey("cooldown");
+    public static final NamespacedKey DROPPED_PLAYER = newKey("dropped-player");
+
+    @Nonnull
+    public static NamespacedKey newKey(@Nonnull String value) {
+        return new NamespacedKey(Cultivation.getInstance(), value);
+    }
+}
