+++ src/main/java/dev/sefiraat/cultivation/managers/TaskManager.java
@@ -1,0 +1,23 @@
+package dev.sefiraat.cultivation.managers;
+
+import com.google.common.base.Preconditions;
+import dev.sefiraat.cultivation.Cultivation;
+
+/**
+ * This class is used to run Runnables from one place
+ */
+public class TaskManager {
+
+    private static TaskManager instance;
+
+    public TaskManager() {
+        Preconditions.checkArgument(instance == null, "Cannot create a new instance of the TaskManager");
+        instance = this;
+        final Cultivation plugin = Cultivation.getInstance();
+
+    }
+
+    public static TaskManager getInstance() {
+        return instance;
+    }
+}
